updating the code to predict ACS
I updated the order of some of the code.  The training model also will not have faminc right now as we have to create categories for the income.  We have started to predict some of the acs data to find the probabiity of a PUMA for households with at least one senior.